move remind later handling to different public class (#40)
